fix(moe_permute): prevent int32 overflow in row*num_experts indexing
When total_zipped_tokens > INT_MAX/num_experts (e.g., >5.6M tokens
with 384 experts), the expression `row * num_experts` overflows int32,
causing silent memory corruption in zipped_expertwise_rowmap access.

Also fix uint32_t overflow in filling_padding_rows_kernel where
`rows * cols` exceeds UINT32_MAX for large hidden dimensions.

All row-stride pointer offsets now use int64_t arithmetic, matching
the existing (int64_t)row * token_length pattern in the same files.

+++ paddle/phi/kernels/gpu/moe_permute_kernel.cu
@@ -265,8 +265,8 @@ __global__ __launch_bounds__(BLOCK_DIM_X) void permute_kernel(
                       __popc((lane_id < 31) ? (mask >> (lane_id + 1)) : 0u);
         }
 
-        zipped_expertwise_rowmap[global_row * num_experts + expert_id] =
-            final_pos;
+        zipped_expertwise_rowmap[(int64_t)global_row * num_experts +
+                                 expert_id] = final_pos;
 #pragma unroll
         for (int k = 0; k < TOPK; k++) {
           if (reg_expert[k] == expert_id) {
